+++ packages/@react-native-windows/telemetry/src/e2etest/telemetry.test.ts
@@ -291,7 +291,7 @@ function getTestCommandProjectInfo(): projectUtils.AppProjectInfo {
     rnwLang: 'cpp',
     usesTS: true,
     usesRNConfig: false,
-    jsEngine: 'Chakra',
+    jsEngine: 'Hermes',
     rnwSource: 'Source',
     dependencies: [
       {