+++ packages/e2e-test-app/.eslintrc.js
@@ -1,9 +1,0 @@
-module.exports = {
-  extends: ['@rnw-scripts'],
-  parserOptions: { tsconfigRootDir: __dirname },
-  globals: {
-    expect: 'readonly',
-    fail: 'readonly',
-    AutomationClient: 'readonly',
-  },
-};