+++ vnext/Microsoft.ReactNative/Modules/AccessibilityInfoModule.cpp
@@ -3,13 +3,7 @@
 
 #include "pch.h"
 #include "AccessibilityInfoModule.h"
-#ifndef USE_FABRIC
-#include <UI.Xaml.Automation.Peers.h>
-#include <UI.Xaml.Controls.h>
-#include <XamlUtils.h>
-#else
 #include <Fabric/Composition/ReactNativeIsland.h>
-#endif
 #include <uiautomationcore.h>
 #include <uiautomationcoreapi.h>
 #include <winrt/Windows.ApplicationModel.DataTransfer.h>
@@ -53,35 +47,6 @@ void AccessibilityInfo::setAccessibilityFocus(double /*reactTag*/) noexcept {
 
 void AccessibilityInfo::announceForAccessibility(std::wstring announcement) noexcept {
   m_context.UIDispatcher().Post([context = m_context, announcement = std::move(announcement)] {
-#ifndef USE_FABRIC
-    // Windows requires a specific element to announce from. Unfortunately the react-native API does not provide a tag
-    // So we need to find something to raise the notification event from.
-    xaml::UIElement element{nullptr};
-
-    if (IsXamlIsland()) {
-      if (auto accessibleRoot =
-              winrt::Microsoft::ReactNative::XamlUIService::GetAccessibleRoot(context.Properties().Handle())) {
-        element = accessibleRoot;
-      } else {
-        return;
-      }
-    } else {
-      element = xaml::Controls::TextBlock();
-    }
-
-    auto peer = xaml::Automation::Peers::FrameworkElementAutomationPeer::FromElement(element);
-
-    if (!peer) {
-      return;
-    }
-
-    winrt::hstring hstr{announcement};
-    peer.RaiseNotificationEvent(
-        xaml::Automation::Peers::AutomationNotificationKind::Other,
-        xaml::Automation::Peers::AutomationNotificationProcessing::ImportantMostRecent,
-        hstr,
-        hstr);
-#else
     if (auto weakIslandWrapper = context.Properties().Get(
             winrt::Microsoft::ReactNative::implementation::ReactNativeIsland::LastFocusedReactNativeIslandProperty())) {
       if (auto weakIsland = weakIslandWrapper.Value()) {
@@ -107,8 +72,6 @@ void AccessibilityInfo::announceForAccessibility(std::wstring announcement) noex
         }
       }
     }
-
-#endif
   });
 }