+++ vnext/Microsoft.ReactNative/ReactHost/React.h
@@ -26,12 +26,13 @@
 
 #include <Shared/IReactRootView.h>
 
-#include <ViewManagerProvider.h>
 #include <winrt/Microsoft.ReactNative.h>
 
-#ifdef USE_FABRIC
 #include <Fabric/Composition/UriImageManager.h>
-#endif
+
+namespace facebook::react::jsinspector_modern {
+class HostTarget;
+} // namespace facebook::react::jsinspector_modern
 
 namespace Mso::React {
 
@@ -50,7 +51,6 @@ enum class LogLevel : int32_t {
 };
 
 enum class JSIEngine : int32_t {
-  Chakra = 0, // Use the JSIExecutorFactory with ChakraRuntime
   Hermes = 1, // Use the JSIExecutorFactory with Hermes
   V8 = 2, // Use the JSIExecutorFactory with V8
 };
@@ -84,17 +84,10 @@ struct IReactInstance : IUnknown {
   virtual ReactInstanceState State() const noexcept = 0;
 
   virtual Mso::React::IReactContext &GetReactContext() const noexcept = 0;
-
-  virtual void AttachMeasuredRootView(
-      facebook::react::IReactRootView *rootView,
-      const winrt::Microsoft::ReactNative::JSValueArgWriter &initialProps,
-      bool useFabric) noexcept = 0;
-  virtual void DetachRootView(facebook::react::IReactRootView *rootView, bool useFabric) noexcept = 0;
 };
 
 MSO_GUID(IReactSettingsSnapshot, "6652bb2e-4c5e-49f7-b642-e817b0fef4de")
 struct IReactSettingsSnapshot : IUnknown {
-  virtual bool UseWebDebugger() const noexcept = 0;
   virtual bool UseFastRefresh() const noexcept = 0;
   virtual bool UseDirectDebugger() const noexcept = 0;
   virtual bool DebuggerBreakOnNextLine() const noexcept = 0;
@@ -120,7 +113,6 @@ struct IReactContext : IUnknown {
   virtual winrt::Microsoft::ReactNative::JsiRuntime JsiRuntime() const noexcept = 0;
   virtual ReactInstanceState State() const noexcept = 0;
   virtual bool IsLoaded() const noexcept = 0;
-  virtual std::shared_ptr<facebook::react::Instance> GetInnerInstance() const noexcept = 0;
   virtual IReactSettingsSnapshot const &SettingsSnapshot() const noexcept = 0;
 };
 
@@ -168,11 +160,6 @@ struct NativeModuleProvider2 {
       std::shared_ptr<facebook::react::MessageQueueThread> const &defaultQueueThread) = 0;
 };
 
-struct ViewManagerProvider2 {
-  virtual std::vector<std::unique_ptr<::Microsoft::ReactNative::IViewManager>> GetViewManagers(
-      Mso::CntPtr<IReactContext> const &reactContext) = 0;
-};
-
 //! A simple struct that describes the basic properties/needs of an SDX. Whenever a new SDX is
 //! getting hosted in React, properties here will be used to construct the SDX.
 struct ReactOptions {
@@ -181,11 +168,8 @@ struct ReactOptions {
   winrt::Microsoft::ReactNative::IReactNotificationService Notifications;
 
   std::shared_ptr<NativeModuleProvider2> ModuleProvider;
-  std::shared_ptr<ViewManagerProvider2> ViewManagerProvider;
   std::shared_ptr<winrt::Microsoft::ReactNative::TurboModulesProvider> TurboModuleProvider;
-#ifdef USE_FABRIC
   std::shared_ptr<winrt::Microsoft::ReactNative::Composition::implementation::UriImageManager> UriImageManager;
-#endif
 
   //! Identity of the SDX. Must uniquely describe the SDX across the installed product.
   std::string Identity;
@@ -216,16 +200,15 @@ struct ReactOptions {
   //! Base path of the SDX. The absolute path of the SDX can be constructed from this and the Identity.
   std::string BundleRootPath;
 
-  //! Javascript Bundles
-  //! This List includes both Platform and User Javascript Bundles
-  //! Bundles are loaded into Javascript engine in the same order
+  //! JavaScript Bundles
+  //! This List includes both Platform and User JavaScript Bundles
+  //! Bundles are loaded into JavaScript engine in the same order
   //! as they are specified in this list.
   std::vector<Mso::CntPtr<IJSBundle>> JSBundles;
 
   //! ReactNative Infrastructure Error
   //! Error types include:
   //! * Any call to Javascript function after Global Exception has been raised
-  //! * Any WebServer error when DeveloperSettings.UseWebDebugger is true
   //! Note: Default callback generates ShipAssert.
   OnErrorCallback OnError{GetDefaultOnErrorHandler()};
 
@@ -237,7 +220,7 @@ struct ReactOptions {
   //! during development to report JavaScript errors to users
   std::shared_ptr<Mso::React::IRedBoxHandler> RedBoxHandler;
 
-  //! Flag to suggest sdx owner's preference on enabling Bytecode caching in Javascript Engine for corresponding SDX.
+  //! Flag to suggest sdx owner's preference on enabling Bytecode caching in JavaScript Engine for corresponding SDX.
   bool EnableBytecode{true};
 
   //! Flag controlling whether the JavaScript engine uses JIT compilation.
@@ -245,11 +228,6 @@ struct ReactOptions {
 
   std::string ByteCodeFileUri;
   bool EnableByteCodeCaching{true};
-
-  //! Enable function nativePerformanceNow.
-  //! Method nativePerformanceNow() returns high resolution time info.
-  //! It is not safe to expose to Custom Function. Add this flag so we can turn it off for Custom Function.
-  bool EnableNativePerformanceNow{true};
 
   ReactDevOptions DeveloperSettings = {};
 
@@ -288,16 +266,6 @@ struct ReactOptions {
   bool UseLiveReload() const noexcept;
   static void SetUseLiveReload(winrt::Microsoft::ReactNative::IReactPropertyBag const &properties, bool value) noexcept;
   static bool UseLiveReload(winrt::Microsoft::ReactNative::IReactPropertyBag const &properties) noexcept;
-
-  //! Should the instance run in a remote environment such as within a browser
-  //! By default, this is using a browser navigated to  http://localhost:8081/debugger-ui served
-  //! by Metro/Haul. Debugging will start as soon as the React Native instance is loaded.
-  void SetUseWebDebugger(bool enabled) noexcept;
-  bool UseWebDebugger() const noexcept;
-  static void SetUseWebDebugger(
-      winrt::Microsoft::ReactNative::IReactPropertyBag const &properties,
-      bool value) noexcept;
-  static bool UseWebDebugger(winrt::Microsoft::ReactNative::IReactPropertyBag const &properties) noexcept;
 
   //! For direct debugging, whether to break on the next line of JavaScript that is executed.
   void SetDebuggerBreakOnNextLine(bool enable) noexcept;
@@ -347,6 +315,9 @@ struct ReactOptions {
   //! The callback is called when IReactInstance is destroyed and must not be used anymore.
   //! It is called from the native queue.
   OnReactInstanceDestroyedCallback OnInstanceDestroyed;
+
+  //! The HostTarget instance for modern inspector integration.
+  facebook::react::jsinspector_modern::HostTarget *InspectorHostTarget;
 };
 
 //! IReactHost manages a ReactNative instance.