+++ vnext/ReactCommon/TEMP_UntilReactCommonUpdate/react/nativemodule/dom/NativeDOM.h
@@ -23,6 +23,15 @@
 
 namespace facebook::react {
 
+using MeasureOnSuccessCallback =
+    SyncCallback<void(double, double, double, double, double, double)>;
+
+using MeasureInWindowOnSuccessCallback =
+    SyncCallback<void(double, double, double, double)>;
+
+using MeasureLayoutOnSuccessCallback =
+    SyncCallback<void(double, double, double, double)>;
+
 class NativeDOM : public NativeDOMCxxSpec<NativeDOM> {
  public:
   NativeDOM(std::shared_ptr<CallInvoker> jsInvoker);
@@ -49,7 +58,9 @@ class NativeDOM : public NativeDOMCxxSpec<NativeDOM> {
       /* rightWidth: */ int,
       /* bottomWidth: */ int,
       /* leftWidth: */ int>
-  getBorderWidth(jsi::Runtime& rt, std::shared_ptr<const ShadowNode> shadowNode);
+  getBorderWidth(
+      jsi::Runtime& rt,
+      std::shared_ptr<const ShadowNode> shadowNode);
 
   std::tuple<
       /* x: */ double,
@@ -66,15 +77,21 @@ class NativeDOM : public NativeDOMCxxSpec<NativeDOM> {
       std::shared_ptr<const ShadowNode> shadowNode);
 
   std::tuple</* scrollLeft: */ double, /* scrollTop: */ double>
-  getScrollPosition(jsi::Runtime& rt, std::shared_ptr<const ShadowNode> shadowNode);
+  getScrollPosition(
+      jsi::Runtime& rt,
+      std::shared_ptr<const ShadowNode> shadowNode);
 
   std::tuple</* scrollWidth: */ int, /* scrollHeight */ int> getScrollSize(
       jsi::Runtime& rt,
       std::shared_ptr<const ShadowNode> shadowNode);
 
-  std::string getTagName(jsi::Runtime& rt, std::shared_ptr<const ShadowNode> shadowNode);
+  std::string getTagName(
+      jsi::Runtime& rt,
+      std::shared_ptr<const ShadowNode> shadowNode);
 
-  std::string getTextContent(jsi::Runtime& rt, std::shared_ptr<const ShadowNode> shadowNode);
+  std::string getTextContent(
+      jsi::Runtime& rt,
+      std::shared_ptr<const ShadowNode> shadowNode);
 
   bool hasPointerCapture(
       jsi::Runtime& rt,
@@ -111,19 +128,19 @@ class NativeDOM : public NativeDOMCxxSpec<NativeDOM> {
   void measure(
       jsi::Runtime& rt,
       std::shared_ptr<const ShadowNode> shadowNode,
-      jsi::Function callback);
+      const MeasureOnSuccessCallback& callback);
 
   void measureInWindow(
       jsi::Runtime& rt,
       std::shared_ptr<const ShadowNode> shadowNode,
-      jsi::Function callback);
+      const MeasureInWindowOnSuccessCallback& callback);
 
   void measureLayout(
       jsi::Runtime& rt,
       std::shared_ptr<const ShadowNode> shadowNode,
       std::shared_ptr<const ShadowNode> relativeToShadowNode,
       jsi::Function onFail,
-      jsi::Function onSuccess);
+      const MeasureLayoutOnSuccessCallback& onSuccess);
 
 // [windows MSVC cant handle backtick character in a preprocessor line
 #pragma mark - Legacy direct manipulation APIs (for ReactNativeElement).