+++ vnext/Shared/Logging.cpp
@@ -7,7 +7,6 @@
 
 #include <cxxreact/ReactMarker.h>
 #include <chrono>
-#include "../Chakra/ChakraPlatform.h"
 
 namespace facebook {
 namespace react {
@@ -20,18 +19,11 @@ void LogHook(RCTLogLevel logLevel, const char *message) {
   g_nativeLogHook(logLevel, message);
 }
 
-static double nativePerformanceNow() {
-  return std::chrono::duration<double, std::milli>(std::chrono::steady_clock::now().time_since_epoch()).count();
-}
-
 void logMarker(const ReactMarker::ReactMarkerId /*id*/, const char * /*tag*/) {}
 } // end anonymous namespace
 
 void InitializeLogging(NativeLoggingHook &&hook) {
   g_nativeLogHook = std::move(hook);
-
-  JSNativeHooks::loggingHook = LogHook;
-  JSNativeHooks::nowHook = nativePerformanceNow;
 
   ReactMarker::logTaggedMarkerImpl = logMarker;
 }