+++ vnext/src-win/Libraries/Components/Pressable/Pressable.windows.js
@@ -71,6 +71,14 @@ type PressableBaseProps = $ReadOnly<{
    */
   disabled?: ?boolean,
 
+  // [Windows
+  /**
+   * When the pressable is pressed it will take focus
+   * Default value: true
+   */
+  focusOnPress?: ?boolean,
+  // Windows]
+
   /**
    * Additional distance outside of this view in which a press is detected.
    */
@@ -120,6 +128,12 @@ type PressableBaseProps = $ReadOnly<{
    * Called when a touch is released before `onPress`.
    */
   onPressOut?: ?(event: GestureResponderEvent) => mixed,
+
+  /**
+   * Whether to prevent any other native components from becoming responder
+   * while this pressable is responder.
+   */
+  blockNativeResponder?: ?boolean,
 
   /**
    * Called after the element loses focus.
@@ -228,6 +242,7 @@ function Pressable({
     'aria-expanded': ariaExpanded,
     'aria-label': ariaLabel,
     'aria-selected': ariaSelected,
+    blockNativeResponder,
     'aria-readonly': ariaReadOnly,
     'aria-multiselectable': ariaMultiselectable, // Windows
     'aria-required': ariaRequired, // Windows
@@ -238,6 +253,7 @@ function Pressable({
     delayLongPress,
     disabled,
     focusable,
+    focusOnPress, // Windows
     hitSlop,
     onBlur,
     onFocus,
@@ -309,6 +325,16 @@ function Pressable({
     hitSlop,
   };
 
+  const onPressWithFocus = React.useCallback(
+    (args: GestureResponderEvent) => {
+      if (focusable !== false && focusOnPress !== false) {
+        viewRef?.current?.focus();
+      }
+      onPress?.(args);
+    },
+    [focusOnPress, onPress, focusable],
+  );
+
   const config = useMemo(
     () => ({
       cancelable,
@@ -325,7 +351,7 @@ function Pressable({
       onHoverIn,
       onHoverOut,
       onLongPress,
-      onPress,
+      onPress: onPressWithFocus,
       onPressIn(event: GestureResponderEvent): void {
         if (android_rippleConfig != null) {
           android_rippleConfig.onPressIn(event);
@@ -350,6 +376,7 @@ function Pressable({
           onPressOut(event);
         }
       },
+      blockNativeResponder,
       // [Windows
       onKeyDown,
       onKeyUp,
@@ -358,6 +385,7 @@ function Pressable({
     [
       android_disableSound,
       android_rippleConfig,
+      blockNativeResponder,
       cancelable,
       delayHoverIn,
       delayHoverOut,
@@ -369,7 +397,7 @@ function Pressable({
       onHoverIn,
       onHoverOut,
       onLongPress,
-      onPress,
+      onPressWithFocus,
       onPressIn,
       onPressMove,
       onPressOut,