+++ vnext/templates/cpp-lib/example/metro.config.js
@@ -1,8 +1,7 @@
 const { getDefaultConfig, mergeConfig } = require('@react-native/metro-config');
 const fs = require('fs');
-const path = require('path');
+const path = require('node:path');
 const escape = require('escape-string-regexp');
-const exclusionList = require('metro-config/src/defaults/exclusionList');
 const pack = require('../package.json');
 
 const root = path.resolve(__dirname, '..');
@@ -33,7 +32,7 @@ const config = {
   // We need to make sure that only one version is loaded for peerDependencies
   // So we block them at the root, and alias them to the versions in example's node_modules
   resolver: {
-    blockList: exclusionList(
+    blocklist: 
       modules.map(
         (m) =>
           new RegExp(`^${escape(path.join(root, 'node_modules', m))}\\/.*$`)
@@ -46,8 +45,7 @@ const config = {
         new RegExp(`${rnwPath}/build/.*`),
         new RegExp(`${rnwPath}/target/.*`),
         /.*\.ProjectImports\.zip/,
-      ])
-    ),
+      ]),
 
     extraNodeModules: modules.reduce((acc, name) => {
       acc[name] = path.join(__dirname, 'node_modules', name);